projects/extractedTextV2/requirements.txt
paddleocr
paddlepaddle
opencv-python
matplotlib
pillow
